+++ next.config.mjs
@@ -1,4 +1,8 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+  env: {
+    openaiKey: process.env.OPENAI_API_KEY, // pulls from .env file
+  },
+};
 
 export default nextConfig;